(split)bug: 50455
Switch mailbox and mailbox group IDs back to ints.

http://bugzilla.zimbra.com/show_bug.cgi?id=50455

Copied from Perforce
 Change: 252986

+++ src/java/com/zimbra/cs/dav/caldav/TimeRange.java
@@ -113,7 +113,7 @@ public class TimeRange {
         return mEnd;
     }
 
-    public boolean matches(long mboxId, int itemId, long apptRangeStart, long apptRangeEnd) {
+    public boolean matches(int mboxId, int itemId, long apptRangeStart, long apptRangeEnd) {
         // it matches if the range of the appointment is completely contained in the requested range.
         if (apptRangeStart >= mStart && apptRangeEnd <= mEnd)
             return true;
