+++ src/java/com/zimbra/cs/db/DbImapMessage.java
@@ -445,7 +445,7 @@ public class DbImapMessage {
     }
 
 
-    public static String getTableName(long mailboxId, long groupId) {
+    public static String getTableName(int mailboxId, int groupId) {
         return DbMailbox.qualifyTableName(groupId, TABLE_IMAP_MESSAGE);
     }
