+++ src/java/com/zimbra/cs/mailbox/CalendarItem.java
@@ -136,7 +136,7 @@ public abstract class CalendarItem extends MailItem implements ScheduledTaskResu
     public TimeZoneMap getTimeZoneMap() { return mTzMap; }
 
     public static class CalendarMetadata {
-        public long mailboxId;
+        public int mailboxId;
         public int itemId;
         public String uid;
         public int mod_metadata;
@@ -144,7 +144,7 @@ public abstract class CalendarItem extends MailItem implements ScheduledTaskResu
         public long start_time;
         public long end_time;
 
-        public CalendarMetadata(long mailboxId, int itemId, String uid, int mod_metadata, int mod_content, long start_time, long end_time) {
+        public CalendarMetadata(int mailboxId, int itemId, String uid, int mod_metadata, int mod_content, long start_time, long end_time) {
             this.mailboxId = mailboxId;
             this.itemId = itemId;
             this.uid = uid;
