+++ src/java/com/zimbra/cs/mailbox/Message.java
@@ -157,7 +157,7 @@ public class Message extends MailItem {
             Invite inv = null;
             Metadata metaInv = meta.getMap(FN_INV, true);
             if (metaInv != null) {
-                long mboxId = mbox.getId();
+                int mboxId = mbox.getId();
                 ICalTimeZone accountTZ = ICalTimeZone.getAccountTimeZone(mbox.getAccount());
                 inv = Invite.decodeMetadata(mboxId, metaInv, null, accountTZ);
             }
