+++ src/java/com/zimbra/cs/mailbox/ScheduledTaskManager.java
@@ -114,14 +114,14 @@ public class ScheduledTaskManager {
         }
     }
     
-    public static ScheduledTask getTask(String className, String taskName, long mailboxId) {
+    public static ScheduledTask getTask(String className, String taskName, int mailboxId) {
         return (ScheduledTask) sScheduler.getTask(getKey(className, taskName, mailboxId));
     }
     
     /**
      * Cancels a persistent task.
      */
-    public static ScheduledTask cancel(String className, String taskName, long mailboxId, boolean mayInterruptIfRunning)
+    public static ScheduledTask cancel(String className, String taskName, int mailboxId, boolean mayInterruptIfRunning)
     throws ServiceException {
         Connection conn = null;
         ScheduledTask task = null;
@@ -145,7 +145,7 @@ public class ScheduledTaskManager {
      * @return the task, or <tt>null</tt> if the task could not be found
      */
     public static ScheduledTask cancel(Connection conn, String className, String taskName,
-                                       long mailboxId, boolean mayInterruptIfRunning)
+                                       int mailboxId, boolean mayInterruptIfRunning)
     throws ServiceException {
         if (conn != null) {
             DbScheduledTask.deleteTask(conn, className, taskName);
@@ -153,7 +153,7 @@ public class ScheduledTaskManager {
         return (ScheduledTask) sScheduler.cancel(getKey(className, taskName, mailboxId), mayInterruptIfRunning);
     }
     
-    private static String getKey(String className, String taskName, long mailboxId) {
+    private static String getKey(String className, String taskName, int mailboxId) {
         StringBuilder sb = new StringBuilder();
         sb.append(className).append(':').append(taskName);
         if (mailboxId > 0) {
