+++ src/java/com/zimbra/cs/mailbox/calendar/cache/FileStore.java
@@ -43,8 +43,8 @@ public class FileStore {
 
     private static final long MAX_CACHE_FILE_LEN = 100 * 1024 * 1024;  // 100MB
 
-    private static File getMailboxDir(long mboxId) {
-        long mdir = mboxId >> MBOX_BITS;
+    private static File getMailboxDir(int mboxId) {
+        int mdir = mboxId >> MBOX_BITS;
         mdir &= MBOX_GROUP_BITS;
 
         StringBuilder sb = new StringBuilder(LC.calendar_cache_directory.value());
@@ -52,10 +52,10 @@ public class FileStore {
         return new File(sb.toString());
     }
 
-    private static File getCalFolderFile(long mboxId, int folderId) {
-        long mdir = mboxId >> MBOX_BITS;
+    private static File getCalFolderFile(int mboxId, int folderId) {
+        int mdir = mboxId >> MBOX_BITS;
         mdir &= MBOX_GROUP_BITS;
-        long fdir = folderId >> FILE_BITS;
+        int fdir = folderId >> FILE_BITS;
         fdir &= FILE_GROUP_BITS;
 
         StringBuilder sb = new StringBuilder(LC.calendar_cache_directory.value());
@@ -64,14 +64,14 @@ public class FileStore {
         return new File(sb.toString());
     }
 
-    static void deleteCalendarData(long mboxId, int folderId)
+    static void deleteCalendarData(int mboxId, int folderId)
     throws ServiceException {
         File file = getCalFolderFile(mboxId, folderId);
         if (file.exists())
             file.delete();
     }
 
-    static void saveCalendarData(long mboxId, CalendarData calData)
+    static void saveCalendarData(int mboxId, CalendarData calData)
     throws ServiceException {
         File file = getCalFolderFile(mboxId, calData.getFolderId());
         try {
@@ -88,7 +88,7 @@ public class FileStore {
         saveToFile(file, encoded);
     }
 
-    static CalendarData loadCalendarData(long mboxId, int folderId, int modSeq)
+    static CalendarData loadCalendarData(int mboxId, int folderId, int modSeq)
     throws ServiceException {
         File file = getCalFolderFile(mboxId, folderId);
         String encoded = loadFromFile(file);
@@ -179,7 +179,7 @@ public class FileStore {
         return null;
     }
 
-    static void removeMailbox(long mboxId) {
+    static void removeMailbox(int mboxId) {
         File dir = getMailboxDir(mboxId);
         try {
             FileUtil.deleteDir(dir);
