+++ src/java/com/zimbra/cs/pop3/Pop3Mailbox.java
@@ -42,7 +42,7 @@ import com.zimbra.cs.mailbox.Message;
 import com.zimbra.cs.mailbox.OperationContext;
 
 class Pop3Mailbox {
-    private long mId; // id of the mailbox
+    private int mId; // id of the mailbox
     private int mNumDeleted; // number of messages deleted
     private long mTotalSize; // raw size from blob store
     private long mDeletedSize; // raw size from blob store
@@ -100,7 +100,7 @@ class Pop3Mailbox {
      *
      * @return the zimbra mailbox id
      */
-    long getId() {
+    int getId() {
         return mId;
     }
