+++ src/java/com/zimbra/cs/redolog/MailboxIdConflictException.java
@@ -21,10 +21,10 @@ public class MailboxIdConflictException extends RedoException {
     private static final long serialVersionUID = -4186818816051395390L;
 
     private String mAccountId;
-    private long mExpectedId;
-    private long mFoundId;
+    private int mExpectedId;
+    private int mFoundId;
 
-    public MailboxIdConflictException(String accountId, long expectedId, long foundId, RedoableOp op) {
+    public MailboxIdConflictException(String accountId, int expectedId, int foundId, RedoableOp op) {
         super("Mailbox ID for account " + accountId + " changed unexpectedly to " + foundId +
               "; expected " + expectedId, op);
         mAccountId = accountId;
@@ -33,6 +33,6 @@ public class MailboxIdConflictException extends RedoException {
     }
 
     public String getAccountId() { return mAccountId; }
-    public long getExpectedId() { return mExpectedId; }
-    public long getFoundId() { return mFoundId; }
+    public int getExpectedId() { return mExpectedId; }
+    public int getFoundId() { return mFoundId; }
 }
